Fix
I messed up, so here is the fix and potential final of shotgun shell

+++ src/main/java/minecrafttransportsimulator/jsondefs/JSONBullet.java
@@ -69,7 +69,7 @@ public class JSONBullet extends AJSONMultiModelProvider {
         public int pellets;
 
         @JSONDescription("How much spread the pellets will have when fired. 0 is no spread, higher values have higher spread.")
-        public int pelletSpreadFactor;
+        public float pelletSpreadFactor;
 
         @JSONDescription("A optional list of effects that this bullet will impart on the entity that it hits.")
         public List<JSONPotionEffect> effects;
